Enable querying lang-server capabilities
We now check the capabilities of the lanaguage server using
the `on_init` callback and then if formatting is not supported,
we setup the fallback formatter with null-ls.
