+++ lua/lsp/null-ls.lua
@@ -1,76 +1,74 @@
 local M = {}
 
-local _, null_ls = pcall(require, "null-ls")
-local utils = require "utils"
-local sources = {}
+local null_ls = require "null-ls"
+local nodejs_local_providers = { "prettier", "prettierd", "prettier_d_slim", "eslint_d", "eslint" }
+local requested_providers = {}
 
-local local_executables = { "prettier", "prettierd", "prettier_d_slim", "eslint_d", "eslint" }
-
-local find_local_exe = function(exe)
-  vim.cmd "let root_dir = FindRootDirectory()"
-  local root_dir = vim.api.nvim_get_var "root_dir"
-  local local_exe = root_dir .. "/node_modules/.bin/" .. exe
-  return local_exe
-end
-
--- https://github.com/jose-elias-alvarez/null-ls.nvim/blob/9b8458bd1648e84169a7e8638091ba15c2f20fc0/doc/BUILTINS.md#eslint
-local get_normalized_exe = function(exe, type)
-  if type == "diagnostics" and exe == "eslint_d" then
-    return "eslint"
-  end
-  return exe
-end
-
-local function setup_ls(exe, type)
-  if utils.has_value(local_executables, exe) then
-    local normalized_exe = get_normalized_exe(exe, type)
-    local smart_executable = null_ls.builtins[type][normalized_exe]
-    local local_executable = find_local_exe(exe)
-    if vim.fn.executable(local_executable) == 1 then
-      smart_executable._opts.command = local_executable
-      table.insert(sources, smart_executable)
-    else
-      if vim.fn.executable(exe) == 1 then
-        smart_executable._opts.command = exe
-        table.insert(sources, smart_executable)
-      end
+local function is_nodejs_provider(provider)
+  for _, local_provider in ipairs(nodejs_local_providers) do
+    if local_provider == provider then
+      return true
     end
+  end
+  return false
+end
+
+local function is_provider_found(provider)
+  -- special case: fallback to "eslint"
+  -- https://github.com/jose-elias-alvarez/null-ls.nvim/blob/9b8458bd1648e84169a7e8638091ba15c2f20fc0/doc/BUILTINS.md#eslint
+  provider = provider == "eslint_d" and "eslint" or provider
+
+  local retval = { is_local = false, path = nil }
+  if vim.fn.executable(provider) == 1 then
+    return false, provider
+  end
+  if is_nodejs_provider(provider) then
+    vim.cmd "let root_dir = FindRootDirectory()"
+    local root_dir = vim.api.nvim_get_var "root_dir"
+    local local_provider = root_dir .. "/node_modules/.bin/" .. provider
+    if vim.fn.executable(local_provider) == 1 then
+      retval.is_local = true
+      retval.path = local_provider
+    end
+  end
+  return retval.is_local, retval.path
+end
+
+local function enable_provider(filetype, provider)
+  local provider_type = "diagnostics" -- this is the most common case
+  if provider == lvim.lang[filetype].formatter.exe then
+    provider_type = "formatting"
+  end
+
+  local is_local, provider_path = is_provider_found(provider)
+
+  if provider_path then
+    if is_local then
+      null_ls.builtins[provider_type][provider]._opts.command = provider_path
+    end
+    table.insert(requested_providers, null_ls.builtins[provider_type][provider])
   else
-    if null_ls.builtins[type][exe] and vim.fn.executable(null_ls.builtins[type][exe]._opts.command) then
-      table.insert(sources, null_ls.builtins[type][exe])
-    end
+    vim.notify(string.format("Unable to find the path for: [%s]", provider), vim.log.levels.WARN)
   end
-  null_ls.register { sources = sources }
 end
 
 -- TODO: for linters and formatters with spaces and '-' replace with '_'
-local function setup(filetype, type)
-  local executables = nil
-  if type == "diagnostics" then
-    executables = lvim.lang[filetype].linters
-  end
-  if type == "formatting" then
-    executables = lvim.lang[filetype].formatter.exe
+function M.setup(filetype)
+  local sources = {}
+  if lvim.lang[filetype].formatter.override_lsp then
+    table.insert(sources, lvim.lang[filetype].formatter.exe)
   end
 
-  if utils.is_table(executables) then
-    for _, exe in pairs(executables) do
-      if exe ~= "" then
-        setup_ls(exe, type)
-      end
+  for _, linter in pairs(lvim.lang[filetype].linters) do
+    table.insert(sources, linter)
+  end
+
+  for _, provider in pairs(sources) do
+    if provider ~= "" then
+      enable_provider(filetype, provider)
     end
   end
-  if utils.is_string(executables) and executables ~= "" then
-    setup_ls(executables, type)
-  end
-end
-
--- TODO: return the formatter if one was registered, then turn off the builtin formatter
-function M.setup(filetype)
-  setup(filetype, "formatting")
-  setup(filetype, "diagnostics")
-  lvim.sources = sources
-  return sources
+  null_ls.register { sources = requested_providers }
 end
 
 return M
